add configurable logging, content match for input files and an option to keep the input files (mostly for debugging)

+++ src/ynab_bank_import/main.py
@@ -2,25 +2,41 @@ from .ynab import YNABStore
 import argparse
 import configparser
 import glob
+import logging
 import os
 import pkg_resources
 
 
-def import_one(importer, input, output):
-    for file in glob.glob(input):
-        print("Input file {}".format(file))
+log = logging.getLogger(__name__)
+
+
+def import_one(importer, input, output,
+               content_match=None,
+               remove_input=False):
+    for filename in glob.glob(input):
+        if (content_match and
+                content_match not in open(filename, encoding='latin1').read()):
+            continue
+
+        log.info("Input file {}".format(filename))
         store = YNABStore(output)
-        print("%s transactions already seen" % len(store.transactions))
-        importer(file, store)
-        os.remove(file)
+        log.info("%s transactions already seen" % len(store.transactions))
+        importer(filename, store)
+        if remove_input:
+            os.remove(filename)
 
         if store.written == 0:
-            print("No new transactions.")
             os.remove(store.output_file)
+            log.info("No new transactions.")
         else:
-            print("Output file: {}".format(store.output_file))
-            print("{} new transactions. {} ignored transactions.".format(
+            log.info("Output file: {}".format(store.output_file))
+            log.info("{} new transactions. {} ignored transactions.".format(
                 store.written, store.ignored))
+
+
+def configure_logging(debug=False):
+    level = logging.DEBUG if debug else logging.INFO
+    logging.basicConfig(level=level, format='%(message)s')
 
 
 def main():
@@ -28,17 +44,26 @@ def main():
     parser.add_argument(
         'configuration', default='accounts.cfg', nargs='?',
         help='Account configuration file. Default: accounts.cfg')
+    parser.add_argument(
+        '-d', '--debug', default=False, action='store_true',
+        help='Enable debug output')
+    parser.add_argument(
+        '-k', '--keep', default=False, action='store_true',
+        help='Keep input files.')
     args = parser.parse_args()
+    configure_logging(debug=args.debug)
 
     config = configparser.ConfigParser()
     config.read(args.configuration)
 
     for section in config.sections():
-        print("Importing {}".format(section))
+        log.info("Importing {}".format(section))
 
         importer = next(pkg_resources.iter_entry_points(
             'ynab_accounts', name=config[section]['type'])).load()
 
         import_one(importer,
                    config[section]['input'],
-                   os.path.join('ynab', section))
+                   os.path.join('ynab', section),
+                   content_match=config[section].get('match'),
+                   remove_input=(not args.keep))
